Update strace in using PEEKDATA for ptrace.

+++ api/src/syscall/task/wait.rs
@@ -93,7 +93,7 @@ pub fn sys_waitpid(pid: i32, exit_code: *mut i32, options: u32) -> AxResult<isiz
         #[cfg(feature = "ptrace")]
         {
             for child in &children {
-                if let Some(status) = starry_ptrace::check_ptrace_stop(child.pid() as i32) {
+                if let Some(status) = starry_ptrace::check_ptrace_stop(child.pid()) {
                     debug!("waitpid: returning ptrace stop for pid={} status=0x{:x}", child.pid(), status);
                     if let Some(exit_code) = exit_code.nullable() {
                         exit_code.vm_write(status)?;
